+++ core/trace/opentelemetry/agent.go
@@ -29,6 +29,7 @@ func StartAgent(c Config) {
 		if len(c.Endpoint) == 0 {
 			return
 		}
+
 		// Just support jaeger now
 		if c.Batcher != "jaeger" {
 			return
@@ -46,9 +47,7 @@ func StartAgent(c Config) {
 			// Always be sure to batch in production.
 			tracesdk.WithBatcher(exp),
 			// Record information about this application in an Resource.
-			tracesdk.WithResource(resource.NewSchemaless(
-				semconv.ServiceNameKey.String(c.Name),
-			)),
+			tracesdk.WithResource(resource.NewSchemaless(semconv.ServiceNameKey.String(c.Name))),
 		)
 
 		otel.SetTracerProvider(tp)
